requirements.txt
apache-airflow
pandas
scikit-learn
gspread
google-auth
oauth2client
sqlalchemy<2.0
seaborn
matplotlib